SinCity Shader + Darksiders2 HUD Removal
1. Added SinCity shader to base post.fx file. (note: best results when
used in conjunction with tonemap shader to reduce gamma)
2. Added an alternate HUD removal PSHash  for Darksiders2 (commented out
by default) that is specific for taking screenshots with the motion blur
effect showing. The default hash removes this effect when HUD disabled.
This alternative hash will remove the main menu for a clean image with
the motion blur still showing.