improved white point calculations.
I still wasn't happy with the way I was dealing with the white point calculations. Didn't seem accurate, and felt redundant with the luminance average setting. I also didn't like tying it to the filmic curve.

Tried normalizing the maximum component color value of the scene, and getting the dot product with the whitepoint input value.
Results are nicer, and more accurate looking than before.

Also updated the comments for these a bit. Tried to be a bit clearer on what the luminance/white point values are actually doing. While keeping it short ;p